SETTINGS PAGE DONE -- WORKING
Allow the user to select the number of iterations of the algorithm
And allow him to launch the algorithm on the website

Due to the fact that we use GNU Octave to launch the algorithm, it's
STRONGLY not recommended to launch the algorithm from the website.

The plateform has been designed so that the script is launched with
MATLAB from task planifier.